+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - [SCopy] Do not copy revisions.
 - [SCexport] Option added to remove revisions when renaming sheets.
+- [SCexport] Fixed view set drop-down error.
 
 #16.2#